# scaffolds/fullstack-app/packages/types/src/index.ts
export type { Database } from './database.types';
export { type ErrorResponse, errorResponseSchema } from './error';
export { type HealthResponse, healthResponseSchema } from './health';
export {
  type CreateNoteInput,
  createNoteInputSchema,
  type ListNotesQuery,
  listNotesQuerySchema,
  type Note,
  type NoteResponse,
  type NotesListResponse,
  noteResponseSchema,
  noteSchema,
  notesListResponseSchema,
} from './notes';